refactor(Providers): Extract all Providers into their own components
